requirements.txt
Flask
python-frontmatter
python-dotenv
requests
google-generativeai
Pillow
